requirements.txt
Django==2.2
django-tastypie==0.14.2
python-dateutil==2.7.3
python-mimeparse==1.6.0
pytz==2018.5
six==1.11.0
psycopg2==2.7.5
